# specs/00010-machine-activation-and-seats/contracts/activation-api.openapi.yaml
openapi: 3.1.0
info:
  title: Machine Activation & Seat Enforcement API
  version: 0.1.0
  description: >
    REST surface for epic E009 (Machine Activation and Seat Enforcement). A licensed application
    activates a machine against an active license (E008), binding a drift-tolerant fingerprint and
    receiving an offline-verifiable, machine-bound credential; seats are enforced race-safely against
    the license's snapshotted `maxActivations` limit; and operators browse / reclaim seats from the
    console. Covers FR-001..024 and SC-001..020.


    TWO AUTH PLANES (this contract deliberately spans both surfaces of the platform; each operation
    declares exactly one).

      * RUNTIME plane — mounted under `/v1/...`, called by the LICENSED APPLICATION (a machine/
        integration). Authenticated by the tenant/machine API key in the `X-API-Key` header, resolved
        to a tenant + capability scopes (see src/server/auth/apikey.ts + app.ts). These operations
        REQUIRE the `activate` scope and fail CLOSED: no resolvable tenant context → `401 unauthorized`;
        a resolvable key that lacks the `activate` scope → `403 forbidden` (FR-002). There is no CSRF on
        this plane — it is header-credential, not cookie, auth. Both runtime activation routes (activate
        and deactivate) are rate-limited to deter seat exhaustion and fingerprint enumeration, keyed per
...
        the limit-exceeded event is audited (FR-013/FR-020).

...


...
    a non-existent one (SC-012).


...

...
servers:
...
paths:
...
  /v1/activations:
    post:
      tags: [activation]
      operationId: activateMachine
      summary: Activate a machine (bind fingerprint, consume a seat, return the machine-bound credential)
      description: >
        Activate a machine against an ACTIVE license and return an offline-verifiable, machine-bound
        credential (FR-001/007; US1). The license is referenced by EITHER its signed `licenseKey`
        (the E001 `LIC1` issued token) OR its `licenseId` (UUID) — exactly one. The client submits a
        `fingerprint` of N salted signal hashes (computed locally; raw identifiers never leave the
        machine, FR-006), a single-use `nonce` (FR-009), and an optional pseudonymous `label`.


        Success paths:

          * `201 Created` — a NEW machine is bound: one seat is consumed and the response carries the
            activation metadata plus the `machineBoundKey` (the re-signed `LIC1` credential that
            verifies offline only on this machine, FR-007).
          * `200 OK` — IDEMPOTENT re-use, returning the ORIGINAL activation (no new seat consumed):
            either the returning machine matched an existing active activation by at least K of N
            signals (drift tolerance, FR-005/US3), or the request replayed the SAME `nonce` for the
            same (license, machine) (FR-009). The `machineBoundKey` is refreshed/returned again.


        Refusals (no seat consumed, no partial activation recorded):

          * `409 seat_limit_reached` — the license is at its `maxActivations` seat limit; details carry
            `{ seatLimit, seatsUsed }` (FR-003/004, SC-002/003).
          * `409 nonce_replayed` — the `nonce` was replayed to forge a DIFFERENT activation (FR-009,
            SC-010).
          * `409 license_not_active` — the license is suspended, revoked, or expired (FR-008, SC-004).
          * `404 license_not_found` — no such license resolves within this tenant (a cross-tenant
            reference also resolves here, FR-015).
          * `400 insufficient_signals` — fewer than the minimum signals needed to form a reliable
            binding (FR-016).
          * `503 signer_unavailable` — the machine-bound credential could not be signed (an E004 signer
            fault); the transaction rolls back FAIL-CLOSED and NO seat is consumed (Plan §Error Handling).


        Requires the `activate` scope; a key lacking it is refused `403 forbidden` (FR-002). The
        surface is rate-limited (FR-013) and every attempt (grant or denial) is audited (FR-014).
      security:
        - apiKey: []
      x-rbac: { requiredScope: activate }
      x-priority: P1
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivateRequest' }
            examples:
              byLicenseKey:
                summary: Activate by signed license key (LIC1 token)
                value:
                  licenseKey: "LIC1.eyJsaWQiOiJlNWY2MDcxOCJ9.aXNzdWVkLXRva2VuLWV4YW1wbGU"
                  fingerprint:
                    signals:
                      - "9f8Xk2QpL0rNvT7bYc4mHs6Jd1WuA3eZoG5iRfB2xKQ"
                      - "2xTt5RbQ9wPmK7vY0nZa8Lc3Jd6HsU1eIoG4fRB2yNw"
                      - "Lp3Qz7Yw1Rk9Vn0Ma8Bc4Hd2Js6Uu5Ei1oGz4fRB2yNw"
                      - "Qm5Rz9Yx2Sk0Wp1Nb7Cc3Hd4Kt8Vv6Fj2pHa5gSC3zOx"
                      - "Rn6Sa0Zy3Tl1Xq2Oc8Dd4Ie5Lu9Ww7Gk3qIb6hTD4aPy"
                  nonce: "b1946ac92492d2347c6235b4d2611184"
                  label: "Design workstation 12"
              byLicenseId:
                summary: Activate by license id (UUID)
                value:
                  licenseId: "e5f60718-3333-4c4d-ae5f-60718293a4b5"
                  fingerprint:
                    signals:
                      - "9f8Xk2QpL0rNvT7bYc4mHs6Jd1WuA3eZoG5iRfB2xKQ"
                      - "2xTt5RbQ9wPmK7vY0nZa8Lc3Jd6HsU1eIoG4fRB2yNw"
                      - "Lp3Qz7Yw1Rk9Vn0Ma8Bc4Hd2Js6Uu5Ei1oGz4fRB2yNw"
                  nonce: "c2a57bd03583e3458d7346c5e3722295"
      responses:
        '201':
          description: >
            A new machine was bound — one seat consumed. Body includes the activation metadata and the
            `machineBoundKey` (offline-verifiable LIC1 credential). The signing key is never returned.
          headers:
            Location:
              required: true
              description: >
                Canonical path of the CREATED activation (`/v1/activations/{activationId}`). REQUIRED on
                the `201` fresh-bind response; deliberately ABSENT on the `200` idempotent re-use response,
                which returns an already-existing activation rather than creating one.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActivationResult' }
              examples:
                bound:
                  value:
                    id: "a7f1c2d3-9a8b-4c7d-8e1f-0a1b2c3d4e5f"
                    licenseId: "e5f60718-3333-4c4d-ae5f-60718293a4b5"
                    machineId: "Zk9Xp2QrL0sNvT7bYc4mHs6Jd1WuA3eZoG5iRfB2xM"
                    status: active
                    activatedAt: "2026-07-13T09:15:00Z"
                    seatLimit: 5
                    seatsUsed: 3
                    label: "Design workstation 12"
                    keyId: "a1b2c3d4-5566-4778-99aa-bbccddeeff00"
                    machineBoundKey: "LIC1.eyJsaWQiOiJlNWY2MDcxOCIsImZwayI6M30.bWFjaGluZS1ib3VuZC1zaWduYXR1cmU"
        '200':
          description: >
            Idempotent re-use — the ORIGINAL activation is returned with no new seat consumed: a
            K-of-N drift match refreshed an existing machine (FR-005), or the same `nonce` was replayed
            for the same (license, machine) (FR-009). Body carries the refreshed `machineBoundKey`.
            On a K-of-N drift re-match the returned identity — `id`, `machineId`, and `activatedAt` — is
            the ORIGINALLY-bound activation's (the stored `machine_id`), NOT the drifted request's; only
            `machineBoundKey` is freshly re-signed and `seatsUsed` reflects the current active count.
            No `Location` header is returned here (the activation already exists; it is not newly created).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActivationResult' }
              examples:
                reused:
                  summary: Returning machine re-used its existing seat
                  value:
                    id: "a7f1c2d3-9a8b-4c7d-8e1f-0a1b2c3d4e5f"
                    licenseId: "e5f60718-3333-4c4d-ae5f-60718293a4b5"
                    machineId: "Zk9Xp2QrL0sNvT7bYc4mHs6Jd1WuA3eZoG5iRfB2xM"
                    status: active
                    activatedAt: "2026-07-13T09:15:00Z"
                    seatLimit: 5
                    seatsUsed: 3
                    label: "Design workstation 12"
                    keyId: "a1b2c3d4-5566-4778-99aa-bbccddeeff00"
                    machineBoundKey: "LIC1.eyJsaWQiOiJlNWY2MDcxOCIsImZwayI6M30.cmVmcmVzaGVkLXNpZ25hdHVyZQ"
        '400':
          description: >
            Either the request body is malformed (`validation_error`) or the fingerprint carries fewer
            than the minimum signals needed to form a reliable binding (`insufficient_signals`, FR-016).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                malformedBody:
                  value:
                    code: validation_error
                    message: "Exactly one of licenseKey or licenseId is required."
                    details: { field: licenseId }
                insufficientSignals:
                  value:
                    code: insufficient_signals
                    message: "Too few fingerprint signals to form a reliable machine binding."
                    details: { signalsProvided: 1, signalsRequired: 3 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/ForbiddenScope' }
        '404': { $ref: '#/components/responses/LicenseNotFound' }
        '409':
          description: >
            The activation is refused and NO seat is consumed: the seat limit is reached
            (`seat_limit_reached`), the `nonce` was replayed to forge a different activation
            (`nonce_replayed`), or the license is not active (`license_not_active`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                seatLimitReached:
                  value:
                    code: seat_limit_reached
                    message: "This license has reached its activation seat limit."
                    details: { seatLimit: 5, seatsUsed: 5 }
                nonceReplayed:
                  value:
                    code: nonce_replayed
                    message: "This activation nonce has already been used for a different activation."
                    details: { reason: replayed_nonce }
                licenseNotActive:
                  value:
                    code: license_not_active
                    message: "The license is not active and cannot be activated."
                    details: { status: suspended }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SignerUnavailable' }
  /v1/activations/{activationId}:
    parameters:
      - $ref: '#/components/parameters/ActivationId'
    delete:
      tags: [activation]
      operationId: deactivateMachine
      summary: Deactivate this machine's own activation (free the seat)
      description: >
        The licensed application deactivates ITS OWN machine, freeing the seat for immediate re-use
        (FR-010; US2). IDEMPOTENT: returns `204` whether the activation was active or already
        deactivated, and never drives the active-seat count below zero (FR-011). A `404` is returned
        ONLY when the `activationId` is unknown within the caller's tenant (a cross-tenant id also
        resolves here, FR-015). Requires the `activate` scope (FR-002); audited (FR-014). No CSRF —
        this is API-key, not cookie, auth.
      security:
        - apiKey: []
      x-rbac: { requiredScope: activate }
      x-priority: P1
      responses:
        '204':
          description: >
            The seat is free — the activation is deactivated, or was already deactivated (idempotent
            no-op). No body either way.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/ForbiddenScope' }
        '404': { $ref: '#/components/responses/ActivationNotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }

  # --- ADMIN plane (console session + RBAC + CSRF) ------------------------------------------------
  /admin/licenses/{licenseId}/activations:
    parameters:
      - $ref: '#/components/parameters/LicenseId'
    get:
      tags: [activation-registry]
      operationId: listLicenseActivations
      summary: Browse a license's activation registry (tenant-scoped)
      description: >
        List a license's activations for operators/admins (FR-012; US4): each entry shows the
        PSEUDONYMOUS `machineId`, `status`, `activatedAt`, `deactivatedAt`, and optional `label`, plus
        a `seatsUsed`-vs-`seatLimit` summary. The machine-bound credential and the raw signal hashes
        are NEVER returned here (SC-011) — only `POST /v1/activations` returns a credential. Bounded to
        a hard cap of 1000 items; not paginated. Tenant-scoped: an unknown or cross-tenant license →
        `404 not_found` (FR-015, SC-012). READ: `viewer` or higher.
      security:
        - sessionCookie: []
      x-rbac: { minRole: viewer }
      x-priority: P1
      responses:
        '200':
          description: The license's activations (pseudonymous; no credential or raw signals) plus seat usage.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActivationRegistry' }
              examples:
                mixed:
                  value:
                    seatsUsed: 1
                    seatLimit: 5
                    activations:
                      - id: "a7f1c2d3-9a8b-4c7d-8e1f-0a1b2c3d4e5f"
                        machineId: "Zk9Xp2QrL0sNvT7bYc4mHs6Jd1WuA3eZoG5iRfB2xM"
                        status: active
                        activatedAt: "2026-07-13T09:15:00Z"
                        deactivatedAt: null
                        label: "Design workstation 12"
                      - id: "b8e2d3c4-1a2b-4d8e-9f2a-1b2c3d4e5f60"
                        machineId: "Ql8Wo1PqK9rMuS6aXb3lGr5Ic0VtZ2dYnF4hQeA1wL"
                        status: deactivated
                        activatedAt: "2026-06-01T12:00:00Z"
                        deactivatedAt: "2026-07-10T08:30:00Z"
                        label: null
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/licenses/{licenseId}/activations/{activationId}/deactivate:
    parameters:
      - $ref: '#/components/parameters/LicenseId'
      - $ref: '#/components/parameters/ActivationId'
    post:
      tags: [activation-registry]
      operationId: deactivateLicenseActivation
      summary: Operator/admin reclaim a machine's seat
      description: >
        An operator/admin reclaims a specific machine's seat from the console, deactivating the
        activation and freeing the seat for immediate re-use (FR-010; US2). IDEMPOTENT: returns `200`
        whether the activation was active or already deactivated, and never drives the seat count below
        zero (FR-011). An unknown or cross-tenant `licenseId`/`activationId` → `404 not_found`
        (FR-015). RBAC: `admin` or higher (a viewer attempt is refused `403 forbidden` and recorded as
        a security event, SC-009); the mutation carries the double-submit `X-CSRF-Token`. Audited
        (FR-014). STATUS DIVERGENCE (by plane convention): the runtime `DELETE /v1/activations/{activationId}`
        returns `204` with no body (a RESTful resource deactivation), whereas this console action returns
        `200` with a `DeactivationResult` so the operator UI can render the resulting status; both are
        idempotent and identical in effect.
      security:
        - sessionCookie: []
      x-rbac: { minRole: admin }
      x-priority: P1
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: The seat is reclaimed — the activation is deactivated (or was already). Idempotent.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeactivationResult' }
              examples:
                reclaimed:
                  value:
                    id: "a7f1c2d3-9a8b-4c7d-8e1f-0a1b2c3d4e5f"
                    status: deactivated
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        RUNTIME plane credential (NOT the human `admin_session` cookie). The tenant/machine API key
        resolved by the server (src/server/auth/apikey.ts + app.ts) to a tenant + capability scopes.
        The `/v1` activation operations REQUIRE the `activate` scope and fail closed — no resolvable
        tenant context → `401 unauthorized`; a key lacking `activate` → `403 forbidden` (FR-002).
        OpenAPI cannot express the scope in the security requirement for an `apiKey` scheme, so the
        required scope is declared per operation via `x-rbac.requiredScope` and enforced server-side.
    sessionCookie:
      type: apiKey
      in: cookie
      name: admin_session
      description: >
        ADMIN plane credential (NOT the machine `X-API-Key`). The opaque `admin_session` cookie
        established by the E005 `POST /admin/auth/login`, set `HttpOnly` + `Secure` +
        `SameSite=Strict`, scoped to `Path=/admin`, with a bounded `Max-Age`. Server-side revocable;
        never a JWT and never returned in any body. RBAC (`owner > admin > viewer`) is enforced via
        each admin operation's `x-rbac.minRole`.
...
  schemas:
    Error:
      type: object
      description: >
        Project-standard structured JSON error `{ code, message, details? }`. `code` is a stable,
        enumerated snake_case machine-readable value; `message` is human-readable; `details` is
        optional structured context whose shape depends on `code`. No field ever carries a raw
        hardware identifier, signing-key material, a signed token, or a nonce; and no field discloses
        another tenant's existence — cross-tenant access resolves to `404` (never `403`).
      required: [code, message]
      properties:
        code:
          type: string
          description: >
            Stable machine-readable error code (snake_case). Enumerated; new codes require a contract
            change. Each refusal reason is a DISTINCT, testable code.
          enum:
            - validation_error      # 400 — malformed body / bad path or query value
            - insufficient_signals  # 400 — fewer than the minimum fingerprint signals (FR-016)
            - unauthorized          # 401 — no resolvable auth (missing/invalid API key or session)
            - forbidden             # 403 — missing `activate` scope, RBAC deny, or CSRF failure (fail-closed)
            - not_found             # 404 — unknown license/activation within the tenant (admin plane + runtime deactivate; cross-tenant resolves here)
            - license_not_found     # 404 — the referenced license does not resolve in this tenant (activate)
            - seat_limit_reached    # 409 — the license is at its maxActivations seat limit (FR-003/004)
            - nonce_replayed        # 409 — the nonce was replayed to forge a different activation (FR-009)
            - license_not_active    # 409 — the license is suspended, revoked, or expired (FR-008)
            - rate_limited          # 429 — runtime activation surface rate limit exceeded (FR-013/FR-020); response carries Retry-After
            - signer_unavailable    # 503 — machine-bound credential could not be signed (E004 signer fault); no seat consumed (Plan §Error Handling)
          example: seat_limit_reached
        message:
          type: string
          description: Human-readable summary. Never contains a raw identifier, key material, token, or nonce.
        details:
          type: object
          additionalProperties: true
          description: >
            Optional structured context whose shape depends on `code`:
            `validation_error` → `{ field, value? }`;
            `insufficient_signals` → `{ signalsProvided, signalsRequired }`;
            `forbidden` → `{ requiredScope }` (runtime) or `{ minRole, role }` / `{ reason: "csrf" }` (admin);
            `not_found` → the missing id (`{ activationId }` or `{ licenseId }`);
            `license_not_found` → `{ licenseId? }` (omitted when referenced by opaque key);
            `seat_limit_reached` → `{ seatLimit, seatsUsed }`;
            `nonce_replayed` → `{ reason: "replayed_nonce" }` (the nonce value is never echoed);
            `license_not_active` → `{ status: "suspended" | "revoked" | "expired" }`;
            `rate_limited` → `{ retryAfterSeconds }` (also surfaced via the `Retry-After` header);
            `signer_unavailable` → `{ retryAfterSeconds? }` (also surfaced via `Retry-After`).
            It never carries a raw identifier, key material, token, or nonce.

    # --- Shared value types ---
...
    ActivateRequest:
      type: object
      description: >
        Activate a machine against a license. Reference the license by EITHER `licenseKey` OR
        `licenseId` (exactly one). Carries the machine `fingerprint`, a single-use `nonce`, and an
        optional pseudonymous `label`.
      additionalProperties: false
      required: [fingerprint, nonce]
      properties:
        licenseKey:
          allOf: [{ $ref: '#/components/schemas/LicenseKeyToken' }]
          description: The license's signed LIC1 token. Provide this OR `licenseId`, not both.
        licenseId:
          type: string
          format: uuid
          description: The license id (UUID) within the caller's tenant. Provide this OR `licenseKey`, not both.
        fingerprint: { $ref: '#/components/schemas/Fingerprint' }
        nonce:
          type: string
          description: >
            Single-use client-supplied nonce for anti-replay (FR-009). MUST be high-entropy
            (>=128-bit single-use, FR-021) and is retained for a bounded replay-rejection window
            (default 24h). A retry with the SAME nonce for the same (license, machine) replays the
            original result (`200`); a nonce replayed to forge a different activation is rejected
            (`409 nonce_replayed`). Never echoed in any response or error.
            UNIQUENESS SCOPE: the nonce is enforced UNIQUE PER TENANT (DB `UNIQUE (tenant_id, nonce)`,
            data-model §6); on a collision the request is REPLAYED (`200`) only when it targets the same
            (license, machine) as the stored activation, otherwise REJECTED (`409 nonce_replayed`) — so
            the per-tenant DB uniqueness and the (license, machine) replay boundary agree.
          minLength: 16
          maxLength: 200
          pattern: '^[A-Za-z0-9_-]+$'
          example: "b1946ac92492d2347c6235b4d2611184"
        label:
          type: string
          description: Optional pseudonymous, human-friendly label for the machine (never a raw identifier).
          maxLength: 200
          example: "Design workstation 12"
      oneOf:
        - required: [licenseKey]
        - required: [licenseId]
    ActivationResult:
      type: object
      description: >
        The activation returned by `POST /v1/activations` (201 fresh or 200 idempotent re-use). Carries
        the activation metadata, the seat summary, and the `machineBoundKey` credential. This is the
        ONLY schema that includes `machineBoundKey`; the signing key and raw signals are never included.
      required: [id, licenseId, machineId, status, activatedAt, seatLimit, seatsUsed, machineBoundKey]
      properties:
        id: { type: string, format: uuid, description: "Server-generated activation UUID, unique within the tenant.", example: "a7f1c2d3-9a8b-4c7d-8e1f-0a1b2c3d4e5f" }
        licenseId: { type: string, format: uuid, description: The license this activation belongs to. }
        machineId:
          type: string
          description: >
            Pseudonymous machine identity — a salted hash derived from the fingerprint. Not a raw
            hardware identifier and not a UUID (FR-006, SC-011).
          example: "Zk9Xp2QrL0sNvT7bYc4mHs6Jd1WuA3eZoG5iRfB2xM"
        status: { $ref: '#/components/schemas/ActivationStatus' }
        activatedAt: { type: string, format: date-time, description: When the activation was first recorded (immutable). }
        seatLimit:
          type: integer
          minimum: 1
          description: The license's snapshotted `maxActivations` seat limit (E008), enforced by this epic (FR-003).
          example: 5
        seatsUsed:
          type: integer
          minimum: 0
          description: Active activations for the license after this call (never exceeds `seatLimit`, FR-003).
          example: 3
        label:
          type: [string, "null"]
          description: The optional pseudonymous label supplied at activation, or null.
          example: "Design workstation 12"
        keyId:
          type: string
          format: uuid
          description: >
            The id of the product signing key that signed the `machineBoundKey` (E004). Opaque identifier
            only — never key material. This is the "opaque signing-key id" FR-018 permits to be returned
            alongside the public credential: informational, used for keyring selection/rotation.
          example: "a1b2c3d4-5566-4778-99aa-bbccddeeff00"
        machineBoundKey: { $ref: '#/components/schemas/MachineBoundKey' }
...
  responses:
...
    Unauthorized:
      description: >
        No resolvable authentication — the runtime `X-API-Key` is missing/invalid, or the admin
        `admin_session` cookie is missing, expired, or revoked.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            noApiKey:
              summary: Runtime plane — missing/invalid API key
              value: { code: unauthorized, message: "Missing or invalid API key." }
            noSession:
              summary: Admin plane — no valid session
              value: { code: unauthorized, message: "Authentication required." }
...